PRO-2997: move coercion.rb out of the reflection directory
File move only, no constant or nesting change, so git records a rename and the
next commit's diff is just the nesting.
